XCF-1958 make sure that firt-run task is not executed during DB upgrade phase
Usually initialization tasks are executed during upgrade phase of sipXconfig startup.
It is too early for first-run task to be executed then (it requires that replication threads are running and functional)
I made FirstRunTask to execute only when it receives application initialization even during 'normal' phase.
There are at least 2 hacks here:
- need to make sure that DataInitializer does not remove first-run tasks
- AdminContext now provides inUpgradePhase method that can be used to determine which phase we are in. The implementation is less than idead (checks for 'tapestry' bean which is only present in normal phase)

I suspect that we will be adjusting this implementation shortly when the watchdog is reworked not to start other sipx services in first-run phase.


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@11330 ab1d8caa-1f67-47f1-9e81-24633a41865c

+++ sipXconfig/neoconf/src/org/sipfoundry/sipxconfig/admin/DataInitializer.java
@@ -21,24 +21,27 @@ import org.springframework.context.ApplicationContextAware;
 public class DataInitializer implements SystemTaskEntryPoint, ApplicationContextAware {
     private AdminContext m_adminContext;
     private ApplicationContext m_app;
-    
+
     public void runSystemTask(String[] args) {
         String[] tasks = m_adminContext.getInitializationTasks();
-        for (int i = 0; i < tasks.length; i++) {
-            initializeData(tasks[i]);
+        for (String task : tasks) {
+            initializeData(task);
         }
-        
-        // unclear exactly why we'd need to ever call exit.  If you find out why,
+
+        // unclear exactly why we'd need to ever call exit. If you find out why,
         // replace this comment w/reason
         if (args.length >= 2 || !"noexit".equals(args[1])) {
             System.exit(0);
         }
     }
-    
+
     void initializeData(String task) {
         InitializationTask event = new InitializationTask(task);
         m_app.publishEvent(event);
-        m_adminContext.deleteInitializationTask(task);        
+        if (!task.equals("first-run")) {
+            // HACK: do not delete first-run task
+            m_adminContext.deleteInitializationTask(task);
+        }
     }
 
     public void setApplicationContext(ApplicationContext applicationContext) {
